+++ src/ResourceObjectsTester.php
@@ -24,7 +24,7 @@ use PHPUnit\Framework\Assert;
 /**
  * Class ResourcesTester
  *
- * @package CloudCreativity\JsonApi
+ * @package CloudCreativity\JsonApi\Testing
  */
 class ResourceObjectsTester extends AbstractTraversableTester
 {